use uint32_t

+++ lib/IonexWriter.h
@@ -17,15 +17,15 @@ class IonexWriter : private boost::noncopyable
 {
     std::ofstream m_outfile; ///< Output file stream
     std::string m_filename; ///< File name
-    std::size_t m_interval;
+    std::uint32_t m_interval;
     bool m_isKlobuchar; ///< Write Klobuchar or regional model
     bool m_isHeaderWritten;
     std::size_t m_tecmapcount; ///< Index of current TEC map
 
 public:
     IonexWriter();
-    IonexWriter(const char *filename, const std::size_t interval, const bool klobuchar = false);
-    IonexWriter(const std::string &filename, const std::size_t interval, const bool klobuchar = false);
+    IonexWriter(const char *filename, const std::uint32_t interval, const bool klobuchar = false);
+    IonexWriter(const std::string &filename, const std::uint32_t interval, const bool klobuchar = false);
     ~IonexWriter();
 
     void open(const char *filename);
